requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
fastapi==0.110.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.23.2
psycopg2==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
uvicorn==0.29.0
